Restructure packages and add licences
