+++ src/main/java/org/gridsuite/ds/server/dto/dynamicmapping/Script.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2021, RTE (http://www.rte-france.com)
+ * Copyright (c) 2022, RTE (http://www.rte-france.com)
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
@@ -15,7 +15,7 @@ import lombok.NoArgsConstructor;
 import java.util.Date;
 
 /**
- * @author Mathieu Scalbert <mathieu.scalbert at rte-france.com>
+ * @author Thang PHAM <quyet-thang.pham at rte-france.com>
  */
 @Data
 @Schema(description = "Script")
