+++ src/main/java/org/gridsuite/ds/server/model/AbstractManuallyAssignedIdentifierEntity.java
@@ -4,7 +4,7 @@
   License, v. 2.0. If a copy of the MPL was not distributed with this
   file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package org.gridsuite.ds.server.repository;
+package org.gridsuite.ds.server.model;
 
 import org.springframework.data.domain.Persistable;
