+++ src/main/java/org/gridsuite/ds/server/service/DynamicSimulationWorkerService.java
@@ -16,10 +16,10 @@ import com.powsybl.network.store.client.NetworkStoreService;
 import com.powsybl.network.store.client.PreloadingStrategy;
 import com.powsybl.timeseries.StringTimeSeries;
 import com.powsybl.timeseries.TimeSeries;
-import org.gridsuite.ds.server.dsl.GroovyCurvesSupplier;
-import org.gridsuite.ds.server.dsl.GroovyEventModelsSupplier;
+import com.powsybl.dynamicsimulation.groovy.GroovyCurvesSupplier;
+import com.powsybl.dynamicsimulation.groovy.GroovyEventModelsSupplier;
 import org.gridsuite.ds.server.dto.DynamicSimulationStatus;
-import org.gridsuite.ds.server.json.DynamicSimulationResultSerializer;
+import com.powsybl.dynamicsimulation.json.DynamicSimulationResultSerializer;
 import org.gridsuite.ds.server.service.notification.NotificationService;
 import org.gridsuite.ds.server.service.timeseries.TimeSeriesService;
 import org.slf4j.Logger;
