+++ src/main/java/org/gridsuite/ds/server/service/dynamicmapping/implementation/DynamicMappingServiceImpl.java
@@ -1,3 +1,9 @@
+/**
+ * Copyright (c) 2022, RTE (http://www.rte-france.com)
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ */
 package org.gridsuite.ds.server.service.dynamicmapping.implementation;
 
 import com.powsybl.commons.PowsyblException;
@@ -12,6 +18,9 @@ import org.springframework.web.util.UriComponentsBuilder;
 
 import javax.annotation.PostConstruct;
 
+/**
+ * @author Thang PHAM <quyet-thang.pham at rte-france.com>
+ */
 @Service
 public class DynamicMappingServiceImpl implements DynamicMappingService {
     private final String baseUri;
