+++ src/main/java/org/gridsuite/ds/server/service/notification/implementation/NotificationServiceImpl.java
@@ -1,3 +1,9 @@
+/**
+ * Copyright (c) 2022, RTE (http://www.rte-france.com)
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ */
 package org.gridsuite.ds.server.service.notification.implementation;
 
 import org.gridsuite.ds.server.service.notification.NotificationService;
@@ -8,6 +14,9 @@ import org.springframework.cloud.stream.function.StreamBridge;
 import org.springframework.messaging.Message;
 import org.springframework.stereotype.Service;
 
+/**
+ * @author Thang PHAM <quyet-thang.pham at rte-france.com>
+ */
 @Service
 public class NotificationServiceImpl implements NotificationService {
     private static final String CATEGORY_BROKER_OUTPUT = NotificationService.class.getName() + ".output-broker-messages";
@@ -16,16 +25,19 @@ public class NotificationServiceImpl implements NotificationService {
     @Autowired
     private StreamBridge publisher;
 
-    @Override public void sendMessage(Message<String> message, String bindingName) {
+    @Override
+    public void sendMessage(Message<String> message, String bindingName) {
         NotificationServiceImpl.OUTPUT_MESSAGE_LOGGER.debug("Sending message : {}", message);
         publisher.send(bindingName, message);
     }
 
-    @Override public void emitRunDynamicSimulationMessage(Message<String> message) {
+    @Override
+    public void emitRunDynamicSimulationMessage(Message<String> message) {
         sendMessage(message, "publishRun-out-0");
     }
 
-    @Override public void emitResultDynamicSimulationMessage(Message<String> message) {
+    @Override
+    public void emitResultDynamicSimulationMessage(Message<String> message) {
         sendMessage(message, "publishResult-out-0");
     }
 }
