+++ src/test/java/org/gridsuite/ds/server/AbstractDynamicSimulationTest.java
@@ -1,3 +1,9 @@
+/**
+ * Copyright (c) 2022, RTE (http://www.rte-france.com)
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ */
 package org.gridsuite.ds.server;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
@@ -33,6 +39,9 @@ import static org.apache.commons.collections4.CollectionUtils.emptyIfNull;
 import static org.gridsuite.ds.server.service.timeseries.TimeSeriesService.*;
 import static org.gridsuite.ds.server.service.parameters.ParametersService.*;
 
+/**
+ * @author Thang PHAM <quyet-thang.pham at rte-france.com>
+ */
 @RunWith(SpringRunner.class)
 @AutoConfigureWebTestClient(timeout = "PT360S")
 @EnableWebFlux
