+++ src/test/resources/com/powsybl/config/test/config.yml
@@ -1,6 +1,5 @@
 dynawaltz:
-#  homeDir: /home/runner/work/dynamic-simulation-server/dynamic-simulation-server/dynawo
-  homeDir: /home/phamquy/Projects/dynawo/dynawo
+  homeDir: /home/runner/work/dynamic-simulation-server/dynamic-simulation-server/dynawo
   debug: true
 
 dynawaltz-default-parameters:
